+++ dio/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Fix `DioErrorType.cancel` in `Interceptors`.
+- Fix wrong encoding of collection query parameters.
 - Fix "unsupported operation" error on web platform.
 
 ## 5.0.1